requirements.txt
google-cloud-secret-manager~=2.2.0
pydantic~=1.7.3
